Create new files and new namespace for public wrapper classes
