+++ test/Microsoft.ML.Core.Tests/Microsoft.ML.Core.Tests.csproj
